Refactor window and view for application state
After considering it, I think sharing the state of a given application
is a better approach going forward.  With one canonical and widely
shared place for the data, there are less methods passing it back and
forth through application layers.  Everyone can access it and modify
it, so the pieces are less dependent on the other and more on the
data.  Which makes sense to me from an application design standpoint,
if you are changing what data is being stored then you are refactoring
the application anyways.