Move connection start and remove data processing
Moving the start method to a more logical place.

Also removed the data processing method.  It was an attempt to keep
the Threads isolated in the program, but with changes to pypaint that
is no longer very feasible.  I may add a base application state class
at some point and encapsulate that sort of logic.