Split off functionality into new controllers
Remove the saving and loading logic from the base controller and put
it in its own controller where the proper methods are enforced by
the abstractmethod decorator.  This simplifies and focuses both.

Following that, I now wrap the connection in a connection controller
that manages it.  The two are tightly coupled, but the pair is more
encapsulated from the rest of the system.  There is a little
dependency on the application state from pypaint at the moment, but
that will be factored out in the next version.