Split controllers into interfaces and components
To avoid eventual problems with multiple inheritance, I split each of
the previous base controllers into a component and a controller class
that functions as an interface.

This way shared arguments, argument orders, and inheritance orders do
not matter.  Each component is managed by the application controller
and the same logic can be accessed through the components instead of
super() calls.  I think it is overall less complexity, the separate
pieces are more explicit.