Fix errors in closing application
Added closing logic for the listener socket so it can be closed when
an application shuts down and prevent the accept call locking up the
shutdown.

Also, catch the error in the receive thread accessing the socket that
has been set to None.

Updated the development status to "Alpha", since I expect lots of
changes as this library is adapted to more of my existing applications.