Add hex_math and main_script modules, and tests
The hex_math module already existed in a couple other modules, so I
pulled it in and removed the auto-generation of functions.  It was
useful and kept the file short, but was not very clear.

The main_script module contains logic for creating a simple, base,
argparser and logger.  I have used a similar design across multiple
projects and they were not specific to any application.